Move machine identity e2e coverage into its own suite
Machines no longer touch the OAuth surface, so their connection and
lifecycle tests move from OAuthMcp to MachineIdentity, leaving the OAuth
suite covering only discovery, the authorize rules, and the CLI device
grant.
